Trait adv2_splitter::codec::AdCodec

pub trait AdCodec {
    fn decode(&self, reader: &mut dyn Read) -> DuplicationResult;
    fn encode(
        &self,
        dupe_info: &DuplicationInfo,
        lua_value: &LuaValue,
        stream: &mut dyn Write
    ) -> Result<(), Box<dyn Error>>;
    fn is_valid_signature(&self, signature: &[u8]) -> bool;
}

Required Methods

fn decode(&self, reader: &mut dyn Read) -> DuplicationResult

Returns a Duplication Result that will contain the parsed info block and

Arguments

reader - A reader to decode and parse into a result struct

Examples

use codec;
let codec = AdCodec5::default();

let input_file = File::open("test.txt").expect("Unable to open file");
let mut reader = BufReader::new(input_file);
let result = codec.decode(input_file)
advanced_dupe::print_result(result);

fn encode(
&self,
dupe_info: &DuplicationInfo,
lua_value: &LuaValue,
stream: &mut dyn Write
) -> Result<(), Box<dyn Error>>

Arguments

dupe_info - Duplication metadata, note the map must be populated

lua_value - This is the LuaValue as a table

stream - This is the output stream to write to, typically a file

Examples

use codec;
let codec = AdCodec5::default();

let input_file = File::open("test.txt").expect("Unable to open file");
let mut reader = BufReader::new(input_file);
let result = codec.decode(input_file)

let output_file = File::create("test2.txt").expect("Unable to open output file");
let mut out_stream = BufWriter::new(output_file);

codec.encode(result.info(), result.value(), out_stream).expect("Unable to encode table");

fn is_valid_signature(&self, signature: &[u8]) -> bool

Returns if the signature byte array is valid or not.

Arguments

signature - Header of the AdvancedDuplicator file